+++ collective/autopublishing/eventhandler.py
@@ -3,12 +3,13 @@ import logging
 from zope.component import ComponentLookupError, getUtility
 
 from plone.registry.interfaces import IRegistry
+from plone import api
 from Products.CMFCore.WorkflowCore import WorkflowException
-from Products.AdvancedQuery import Eq, In, Le
-from Products.CMFCore.utils import getToolByName
 
 from collective.complexrecordsproxy import ComplexRecordsProxy
 from browser.autopublishsettings import IAutopublishSettingsSchema
+from interfaces import IBrowserLayer
+
 
 logger = logging.getLogger('collective.autopublishing')
 
@@ -19,8 +20,53 @@ logger = logging.getLogger('collective.autopublishing')
 # one of the tick events from collective.timedevents
 
 
+def getExpirationDate(obj):
+    # Archetypes
+    try:
+        date = obj.getExpirationDate()
+        return date
+    # Handle dexterity
+    except AttributeError:
+        date = obj.expires()
+        return date
+
+    return None
+
+
+def setExpirationDate(obj, date):
+    obj.setExpirationDate(date)
+
+
+def getEffectiveDate(obj):
+    # Archetypes
+    try:
+        date = obj.getEffectiveDate()
+        return date
+    # Handle dexterity
+    except AttributeError:
+        date = obj.effective
+        return date
+
+    return None
+
+
+def assemble_mail_text(record):
+    messageText = record['header'] + '\n'
+    messageText += "Content types:" + str(record['content_types'])
+    messageText += "\nInitial state:" + str(record['initial_state'])
+    messageText += "\nTransition:" + str(record['transition'])
+    messageText += "\nActions:" + '\n'
+    for action in record['actions']:
+        messageText += "Transition:" + str(action['transition'])
+        messageText += "\nPortal type:" + str(action['portal_type'])
+        messageText += "\nTitle:" + str(action['title'])
+        messageText += "\nUrl:" + str(action['url']) + '\n\n'
+        messageText += '\n\n'
+    return messageText
+
+
 def autopublish_handler(event):
-    catalog = event.context.portal_catalog
+    catalog = api.portal.get_tool(name='portal_catalog')
 
     try:
         settings = getUtility(IRegistry).forInterface(
@@ -28,7 +74,8 @@ def autopublish_handler(event):
             omit=('publish_actions', 'retract_actions'),
             factory=ComplexRecordsProxy)
     except (ComponentLookupError, KeyError):
-        logger.info('The product needs to be installed. No settings in the registry.')
+        logger.info('The product needs to be installed. '
+                    'No settings in the registry.')
         return
 
     if 'enableAutopublishing' not in catalog.indexes():
@@ -46,105 +93,50 @@ def autopublish_handler(event):
 
         messageText += 'Publish actions:\n'
         for record in p_result:
-            messageText += record['header'] + '\n'
-            messageText += "Content types:" + str(record['content_types']) + '\n'
-            messageText += "Initial state:" + str(record['initial_state']) + '\n'
-            messageText += "Transition:" + str(record['transition']) + '\n'
-            messageText += "Date index/method:" + str(record['date_index_method']) + '\n'
-            messageText += "Actions:" + '\n'
-            for action in record['actions']:
-                messageText += "Transition:" + str(action['transition']) + '\n'
-                messageText += "Portal type:" + str(action['portal_type']) + '\n'
-                messageText += "Title:" + str(action['title']) + '\n'
-                messageText += "Url:" + str(action['url']) + '\n\n'
-            messageText += '\n\n'
+            messageText += assemble_mail_text(record)
 
         messageText += '\n\nRetract actions:\n'
         for record in r_result:
-            messageText += record['header'] + '\n'
-            messageText += "Content types:" + str(record['content_types']) + '\n'
-            messageText += "Initial state:" + str(record['initial_state']) + '\n'
-            messageText += "Transition:" + str(record['transition']) + '\n'
-            messageText += "Date index/method:" + str(record['date_index_method']) + '\n'
-            messageText += "Actions:" + '\n'
-            for action in record['actions']:
-                messageText += "Transition:" + str(action['transition']) + '\n'
-                messageText += "Portal type:" + str(action['portal_type']) + '\n'
-                messageText += "Title:" + str(action['title']) + '\n'
-                messageText += "Url:" + str(action['url']) + '\n\n'
-            messageText += '\n\n'
+            messageText += assemble_mail_text(record)
 
-        email_addresses = settings.email_log
-        mh = getToolByName(event.context, 'MailHost')
-        portal = getToolByName(event.context, 'portal_url').getPortalObject()
-        mh.send(messageText,
-                mto=email_addresses,
-                mfrom=portal.getProperty('email_from_address'),
-                subject='Autopublishing results',
-                encode=None,
-                immediate=False,
-                charset='utf8',
-                msg_type=None)
+        api.portal.send_email(
+            body=messageText,
+            recipient=settings.email_log,
+            subject='Autopublishing results')
 
 
 def handle_publishing(context, settings, dry_run=True, log=True):
     '''
     '''
-    catalog = context.portal_catalog
-    wf = context.portal_workflow
+    catalog = api.portal.get_tool(name='portal_catalog')
+    wf = api.portal.get_tool(name='portal_workflow')
     now = context.ZopeTime()
 
     actions = settings.publish_actions
     action_taken = False
-    audit = ''
     audit = []
     for a in actions:
         audit_record = {}
 
-        date_index = 'effective'
-        date_method = 'getEffectiveDate'
-        date_index_value = a.date_index
-        if date_index_value:
-            if '|' in date_index_value:
-                items = date_index_value.split('|')
-                _date_index = items[0]
-                _date_method = items[1]
-            else:
-                _date_index = date_index_value
-                _date_method = date_index_value
-            if _date_index in catalog.indexes():
-                date_index = _date_index
-                date_method = _date_method
-            else:
-                logger.warn(
-                    "date index does not exist: %s" % (str(_date_index)))
-                continue
+        date_index = getattr(a, 'date_index', 'effective')
 
         audit_record['header'] = 'Actions triggered by "%s"' % str(date_index)
         audit_record['content_types'] = str(a.portal_types)
         audit_record['initial_state'] = str(a.initial_state)
         audit_record['transition'] = str(a.transition)
-        audit_record['date_index_method'] = (str(date_index) + '/' +
-                                             str(date_method))
         audit_record['actions'] = []
 
-        query = (Eq('review_state', a.initial_state)
-                 & Le(date_index, now)
-                 & Eq('enableAutopublishing', True)
-                 & In('portal_type', a.portal_types))
+        query = {'review_state': a.initial_state,
+                 date_index: {'query': now, 'range': 'max'},
+                 'enableAutopublishing': True,
+                 'portal_type': a.portal_types}
 
-        brains = catalog.evalAdvancedQuery(query)
+        brains = catalog(**query)
         affected = 0
         total = 0
         for brain in brains:
-            o = brain.getObject()
-            try:
-                eff_date = getattr(o, date_method)()
-            except AttributeError:
-                logger.warn(
-                    "date field does not exist: %s" % (str(date_method)))
-                continue
-            exp_date = o.getExpirationDate()
+            eff_date = brain.effective
+            exp_date = brain.expires
             # The dates in the indexes are always set!
             # So unless we test for actual dates on the
             # objects, objects with no EffectiveDate are also published.
@@ -169,6 +161,7 @@ def handle_publishing(context, settings, dry_run=True, log=True):
                 action_taken = True
                 if not dry_run:
                     try:
+                        o = brain.getObject()
                         wf.doActionFor(o, a.transition)
                         o.reindexObject()
                         affected += 1
@@ -194,8 +187,8 @@ def handle_publishing(context, settings, dry_run=True, log=True):
 def handle_retracting(context, settings, dry_run=True, log=True):
     '''
     '''
-    catalog = context.portal_catalog
-    wf = context.portal_workflow
+    catalog = api.portal.get_tool(name='portal_catalog')
+    wf = api.portal.get_tool(name='portal_workflow')
     now = context.ZopeTime()
 
     actions = settings.retract_actions
@@ -204,50 +197,25 @@ def handle_retracting(context, settings, dry_run=True, log=True):
     for a in actions:
         audit_record = {}
 
-        date_index = 'expires'
-        date_method = 'getExpirationDate'
-        date_index_value = a.date_index
-        if date_index_value:
-            if '|' in date_index_value:
-                items = date_index_value.split('|')
-                _date_index = items[0]
-                _date_method = items[1]
-            else:
-                _date_index = date_index_value
-                _date_method = date_index_value
-            if _date_index in catalog.indexes():
-                date_index = _date_index
-                date_method = _date_method
-            else:
-                logger.warn(
-                    "date index does not exist: %s" % (str(_date_index)))
-                continue
+        date_index = getattr(a, 'date_index', 'effective')
 
         audit_record['header'] = 'Actions triggered by "%s"' % str(date_index)
         audit_record['content_types'] = str(a.portal_types)
         audit_record['initial_state'] = str(a.initial_state)
         audit_record['transition'] = str(a.transition)
-        audit_record['date_index_method'] = (str(date_index) + '/' +
-                                             str(date_method))
         audit_record['actions'] = []
 
-        query = (In('review_state', a.initial_state)
-                 & Le(date_index, now)
-                 & Eq('enableAutopublishing', True)
-                 & In('portal_type', a.portal_types))
-
-        brains = catalog.evalAdvancedQuery(query)
+        query = {'review_state': a.initial_state,
+                 date_index: {'query': now, 'range': 'max'},
+                 'enableAutopublishing': True,
+                 'portal_type': a.portal_types}
+        brains = catalog(**query)
 
         affected = 0
         total = 0
         for brain in brains:
-            o = brain.getObject()
-            try:
-                exp_date = getattr(o, date_method)()
-            except AttributeError:
-                logger.warn(
-                    "date field does not exist: %s" % (str(date_method)))
-                continue
+
+            exp_date = brain.expires
             # The dates in the indexes are always set.
             # So we need to test on the objects if the dates
             # are actually set.
@@ -266,6 +234,7 @@ def handle_retracting(context, settings, dry_run=True, log=True):
                 action_taken = True
                 if not dry_run:
                     try:
+                        o = brain.getObject()
                         wf.doActionFor(o, a.transition)
                         o.reindexObject()
                         affected += 1
@@ -273,14 +242,12 @@ def handle_retracting(context, settings, dry_run=True, log=True):
                         logger.info(
                             """The state '%s' of the workflow associated with the
                                object at '%s' does not provide the '%s' action
-                            """ % (brain.review_state,
-                                   o.getURL()),
-                            str(a.transition))
+                            """, brain.review_state, brain.getPath(), str(a.transition))  # noqa
                 audit_record['actions'].append(audit_action)
 
         if log:
-            logger.info("""Ran collective.autopublishing (retract): %d objects found, %d affected
-                    """ % (total, affected))
+            logger.info("Ran collective.autopublishing (retract): %d objects found, %d affected",  # noqa
+                        total, affected)
         audit.append(audit_record)
     if action_taken:
         return audit
@@ -295,18 +262,15 @@ def transition_handler(event):
     # the editors responsibility (the effective date should
     # be checked, but it is a common mistake
     # to expect withdrawal to private to be final.
+    if not IBrowserLayer.providedBy(event.object.REQUEST):
+        return
     if not event.transition:
         return
     if not event.object:
         return
-    # Ignore items which don't have getExpirationDate
-    # TODO: Fix this properly!
-    if not hasattr(event.object, 'getExpirationDate'):
-        return
     now = event.object.ZopeTime()
     # todo: make states into a setting
     if event.transition.id in ['retract', 'reject']:
-        overwrite = False
         try:
             settings = getUtility(IRegistry).forInterface(
                 IAutopublishSettingsSchema,
@@ -316,12 +280,10 @@ def transition_handler(event):
             logger.info('The product needs to be installed.'
                         ' No settings in the registry.')
             settings = None
-        if settings and settings.overwrite_expiration_on_retract:
-            overwrite = True
-        if event.object.getExpirationDate() is None or overwrite:
-            event.object.setExpirationDate(now)
+        overwrite = settings and settings.overwrite_expiration_on_retract
+        if getExpirationDate(event.object) is None or overwrite:
+            setExpirationDate(event.object, now)
     if event.transition.id in ['publish']:
-        overwrite = False
         try:
             settings = getUtility(IRegistry).forInterface(
                 IAutopublishSettingsSchema,
@@ -332,8 +294,6 @@ def transition_handler(event):
                         ' No settings in the registry.')
             settings = None
         if settings and settings.clear_expiration_on_publish:
-            overwrite = True
-        if overwrite:
-            if event.object.getExpirationDate() < now:
+            if getExpirationDate(event.object) < now:
                 # to avoid immediate re-retraction
-                event.object.setExpirationDate(None)
+                setExpirationDate(event.object, None)