+++ collective/autopublishing/interfaces.py
@@ -1,5 +1,9 @@
+# -*- coding: utf-8 -*-
+
 from zope.interface import Interface
+
 
 class IBrowserLayer(Interface):
     """ browser layer for the package  """
 
+# EOF