+++ collective/autopublishing/locales/update.sh
@@ -14,11 +14,11 @@ ERRORS=`find . -name "*pt" | xargs i18ndude find-untranslated | grep -e '^-ERROR
 FATAL=`find . -name "*pt"  | xargs i18ndude find-untranslated | grep -e '^-FATAL' | wc -l`
 
 echo
-echo "There are $WARNINGS warnings \(possibly missing i18n markup\)"
-echo "There are $ERRORS errors \(almost definitely missing i18n markup\)"
-echo "There are $FATAL fatal errors \(template could not be parsed, eg. if it\'s not html\)"
-echo "For more details, run \'find . -name \"\*pt\" \| xargs i18ndude find-untranslated\' or"
-echo "Look the rebuild i18n log generate for this script called \'rebuild_i18n.log\' on locales dir"
+echo "There are $WARNINGS warnings (possibly missing i18n markup)"
+echo "There are $ERRORS errors (almost definitely missing i18n markup)"
+echo "There are $FATAL fatal errors (template could not be parsed, eg. if its not html)"
+echo "For more details, run 'find . -name \"*pt\" | xargs i18ndude find-untranslated' or"
+echo "Look the rebuild i18n log generate for this script called 'rebuild_i18n.log' on locales dir"
 
 rm ./rebuild_i18n.log