+++ collective/autopublishing/tests/test_autopublishing.py
@@ -1,22 +1,24 @@
-import unittest
+# -*- coding: utf-8 -*-
+
 import doctest
 import interlude
+from plone.testing import layered
 
-from Testing import ZopeTestCase as ztc
+from collective.autopublishing.tests.layer import C_AUTOPUBLISHING_LAYER
 
-from collective.autopublishing.tests.layer import MyFunctionalTestCase
 
-optionflags = (doctest.NORMALIZE_WHITESPACE|
-               doctest.ELLIPSIS|
-               doctest.REPORT_NDIFF)
+optionflags = (doctest.NORMALIZE_WHITESPACE |
+               doctest.ELLIPSIS |
+               doctest.REPORT_NDIFF |
+               doctest.REPORT_ONLY_FIRST_FAILURE)
+
 
 def test_suite():
-    suite = ztc.FunctionalDocFileSuite(
+    suite = layered(doctest.DocFileSuite(
         'test_autopublishing.txt',
         optionflags=optionflags,
-        test_class=MyFunctionalTestCase,
-        globs=dict(interact=interlude.interact))
+        globs=dict(interact=interlude.interact)),
+        layer=C_AUTOPUBLISHING_LAYER)
     return suite
 
-if __name__ == '__main__':
-    unittest.main(defaultTest='test_suite')
+# EOF